chore: 테스트 유틸리티 파일 mockUtils.js를 __mocks__ 폴더로 분리
